fghs

+++ src/com/company/Person.java
@@ -4,10 +4,10 @@ package com.company;
  * Created by fs279 on 4/19/17.
  */
 public class Person {
-    public Person(int p){
+    public Person(double p){
         person = p;
     }
-    public int person;
+    public double person;
     public double score1 = 0;
     public double score2 = 0;
 }
